Change frontend structure a bit
